split up the kernel density scripts from the featurs scripts
