+++ features.py
@@ -18,9 +18,11 @@ class Features:
     @staticmethod
     def getFeatures(slice, mels=256):
         y, sr = tuple(slice)
-        y = effects.percussive(y)
+        # y = effects.percussive(y)
         S = feature.melspectrogram(y, sr=sr, n_mels=mels)
         log_S = logamplitude(S, ref_power=np.max)
+        display_spec(log_S, sr)
+
         detector = CENSURE()
         detector.detect(log_S)
         kp = detector.keypoints
@@ -28,7 +30,6 @@ class Features:
 
 def feat_censure(slice):
     from skimage.feature import CENSURE
-    import matplotlib.pyplot as plt
 
     y, sr = tuple(slice)
     y = librosa.effects.percussive(y)
@@ -39,51 +40,10 @@ def feat_censure(slice):
 
     detector.detect(log_S)
     kp = detector.keypoints
-    xx, yy, zz = kd_feature(kp, 5.0, metric='manhattan')
-
-    plt.pcolormesh(xx, yy, zz)#, cmap=plt.cm.gist_heat)
-    plt.scatter(x=kp[:, 1], y=kp[:, 0], s=2 ** detector.scales, facecolor='white', alpha=.5)
-    plt.axis('off')
-    plt.show()
-
-def kd_feature(scatter, bandwidth, xbins=100j, ybins=100j, **kwargs):
-    """Build 2D kernel density estimate (KDE)."""
-    from sklearn.neighbors import KernelDensity
-
-    x = scatter[:, 1]; y = scatter[:, 0]
-
-    # create grid of sample locations (default: 100x100)
-    xx, yy = np.mgrid[x.min():x.max():xbins,
-             y.min():y.max():ybins]
-
-    xy_sample = np.vstack([yy.ravel(), xx.ravel()]).T
-    xy_train = np.vstack([y, x]).T
-
-    kde_skl = KernelDensity(bandwidth=bandwidth, **kwargs)
-    kde_skl.fit(xy_train)
-
-    # score_samples() returns the log-likelihood of the samples
-    z = np.exp(kde_skl.score_samples(xy_sample))
-    return xx, yy, np.reshape(z, xx.shape)
-
+    return kp
 
 def display_spec(S, sr):
     plt.figure(figsize=(12, 4))
     librosa.display.specshow(S, sr=sr, x_axis='time', y_axis='mel')
     plt.tight_layout()
     plt.show()
-
-def main():
-    song_folder = os.path.join(os.getcwd(), 'audio/')
-    song_path = os.path.join(song_folder, 'smack_my_b.mp3')
-
-    y, sr = librosa.load(path=song_path, offset=109.12, duration=3.529*2)
-    y = librosa.effects.percussive(y)
-    S = librosa.feature.melspectrogram(y, sr=sr, n_mels=256)
-    mel_slice = librosa.logamplitude(S, ref_power=np.max)
-    display_spec(mel_slice, sr)
-
-    feat_censure(mel_slice)
-
-if __name__ == '__main__':
-    main()
